requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
git+https://github.com/cIoudlab-urv/lithops-Obi-FaaS.git
numpy
Click
tabulate
pandas
PyYAML
python-dateutil
pika
tqdm
lxml
tblib
docker
boto3
requests
seaborn
paramiko
matplotlib
kubernetes
ibm-cos-sdk
ibm-code-engine-sdk
redis
ibm-vpc
cloudpickle
ps-mem
psutil
optuna
scikit-learn
grpcio
grpcio-tools
utils
plotly
kaleido